requirements.txt
streamlit>=1.28.0
Pillow>=9.0.0
crewai>=0.1.0
crewai-tools[mcp]>=0.1.0
requests>=2.25.0
python-dotenv>=0.19.0
groq>=0.1.0
openai>=1.0.0
anthropic>=0.3.0 